requirements.txt
paho-mqtt
python-dotenv
requests
winotify
pystray
pillow
